Move chains to a separate file
